Satis Track Version 1.1.2 - Production Build
This is the full production build of Satis Track. Changes in folder structure was made in order to run the build command using babel and webpack.

Major changes were made to how dates are used on the application in order to account for users from different timezones. This was done by finding the user's timezone and using that information to generate dates to be saved to the database. Email reminders also now work based on the user's timezone.